Extracted JSON persistence functionality
